refactor(control): rename the file for better understanding
